deployment/requirements.txt
streamlit
pandas
matplotlib
plotly
numpy
seaborn
scipy
sklearn
joblib
Pillow